move notifications to separate file so processTransaction and createTransaction actions able to use it
