+++ src/logic/safe/transactions/index.js
@@ -4,3 +4,4 @@ export * from './send'
 export * from './safeBlockchainOperations'
 export * from './safeTxSignerEIP712'
 export * from './txHistory'
+export * from './notifications'
